+++ lib/configuration.ex
@@ -5,14 +5,14 @@ defmodule ExInsights.Configuration do
 
   @doc """
   Reads configuration related to the ex_insights app.
-  
+
   Provides support for `{:system, "VAR_NAME"}` configuration.
   Intended for internal use.
   """
   @spec get_value(atom, default :: term) :: any
   def get_value(key, default \\ nil) do
-     Application.get_env(:ex_insights, key, default)
-     |> return_value()
+    Application.get_env(:ex_insights, key, default)
+    |> return_value()
   end
 
   defp return_value({:system, key}) when is_binary(key) do
@@ -20,5 +20,4 @@ defmodule ExInsights.Configuration do
   end
 
   defp return_value(val), do: val
-
 end