+++ lib/data/payload.ex
@@ -10,39 +10,71 @@ defmodule ExInsights.Data.Payload do
   @doc """
   Create custom event payload.
   """
-  def create_event_payload(name, properties, measurements) do
+  def create_event_payload(name, properties, measurements, instrumentation_key \\ nil) do
     %{
       name: name,
       properties: properties,
       measurements: measurements
     }
-    |> create_payload("Event")
+    |> create_payload("Event", instrumentation_key)
   end
 
   @doc """
   Create custom trace payload.
   """
-  def create_trace_payload(message, severity_level, properties) do
+  def create_trace_payload(message, severity_level, properties, instrumentation_key \\ nil) do
     %{
       message: message,
       properties: properties,
       severityLevel: Utils.convert(severity_level)
     }
-    |> create_payload("Message")
+    |> create_payload("Message", instrumentation_key)
   end
 
   @doc """
   Create custom exception payload.
   """
-  def create_exception_payload(%{__exception__: true, __struct__: type_name, message: message}, stack_trace, handle_at, properties, measurements) do
-    do_create_exception_payload(inspect(type_name), message, stack_trace, handle_at, properties, measurements)
+  def create_exception_payload(
+        %{__exception__: true, __struct__: type_name, message: message},
+        stack_trace,
+        handle_at,
+        properties,
+        measurements,
+        instrumentation_key \\ nil
+      ) do
+    do_create_exception_payload(
+      inspect(type_name),
+      message,
+      stack_trace,
+      handle_at,
+      properties,
+      measurements,
+      instrumentation_key
+    )
   end
 
-  def create_exception_payload(exception, stack_trace, handle_at, properties, measurements) when is_binary(exception) do
-    do_create_exception_payload("Thrown", exception, stack_trace, handle_at, properties, measurements)
+  def create_exception_payload(exception, stack_trace, handle_at, properties, measurements, instrumentation_key \\ nil)
+      when is_binary(exception) do
+    do_create_exception_payload(
+      "Thrown",
+      exception,
+      stack_trace,
+      handle_at,
+      properties,
+      measurements,
+      instrumentation_key
+    )
   end
 
-  defp do_create_exception_payload(type_name, message, stack_trace, handle_at, properties, measurements) do
+  defp do_create_exception_payload(
+         type_name,
+         message,
+         stack_trace,
+         handle_at,
+         properties,
+         measurements,
+         instrumentation_key
+       ) do
     %{
       handledAt: handle_at || "unhandled",
       exceptions: [
@@ -57,30 +89,40 @@ defmodule ExInsights.Data.Payload do
       properties: properties,
       measurements: measurements
     }
-    |> create_payload("Exception")
+    |> create_payload("Exception", instrumentation_key)
   end
 
   @doc """
   Create custom metric payload.
   """
-  def create_metric_payload(name, value, properties) do
+  def create_metric_payload(name, value, properties, instrumentation_key \\ nil) do
     %{
       metrics: [
         %{
           name: name,
           value: value,
-          kind: 0 # Measurement = 0, Aggregation = 1
+          # Measurement = 0, Aggregation = 1
+          kind: 0
         }
       ],
       properties: properties
     }
-    |> create_payload("Metric")
+    |> create_payload("Metric", instrumentation_key)
   end
 
   @doc """
   Create custom dependency payload.
   """
-  def create_dependency_payload(name, command_name, elapsed_time_ms, success, dependency_type_name, target, properties) do
+  def create_dependency_payload(
+        name,
+        command_name,
+        elapsed_time_ms,
+        success,
+        dependency_type_name,
+        target,
+        properties,
+        instrumentation_key \\ nil
+      ) do
     %{
       name: name,
       data: command_name,
@@ -90,24 +132,29 @@ defmodule ExInsights.Data.Payload do
       type: dependency_type_name,
       properties: properties
     }
-    |> create_payload("RemoteDependency")
+    |> create_payload("RemoteDependency", instrumentation_key)
   end
 
   @doc """
   Create request payload
   """
-  def create_request_payload(name, url, source \\ nil, elapsed_time_ms, result_code, success \\ nil, properties, measurements) do
-    success =
-      if nil == success do
-        result_code >= 200 and result_code < 300
-      else
-        success
-      end
-
+  def create_request_payload(
+        name,
+        url,
+        source,
+        elapsed_time_ms,
+        result_code,
+        success,
+        properties,
+        measurements,
+        id \\ nil,
+        instrumentation_key \\ nil
+      ) do
+     id = if (id == nil), do: Base.encode16(<<:rand.uniform(438_964_124)::size(32)>>), else: id
     %{
       name: name,
       url: url,
-      id: Base.encode16(<<:rand.uniform(438964124) :: size(32)>>),
+      id: id,
       source: source,
       duration: Utils.ms_to_timespan(elapsed_time_ms),
       responseCode: result_code,
@@ -115,11 +162,17 @@ defmodule ExInsights.Data.Payload do
       properties: properties,
       measurements: measurements
     }
-    |> create_payload("Request")
+    |> create_payload("Request", instrumentation_key)
   end
 
-  defp create_payload(data, type) do
+  defp create_payload(data, type, instrumentation_key) do
+    instrumentation_key = if (instrumentation_key == nil), do: Conf.get_value(:instrumentation_key), else: instrumentation_key
     data
-    |> Envelope.create(type, DateTime.utc_now(), Conf.get_value(:instrumentation_key), Envelope.get_tags())
+    |> Envelope.create(
+      type,
+      DateTime.utc_now(),
+      instrumentation_key,
+      Envelope.get_tags()
+    )
   end
 end